+++ evo1/seq/__init__.py
@@ -1,7 +1,8 @@
 from evo1.seq.edel_vale import Edel1
-from evo1.seq.overworld import OverworldToMeadow
+from evo1.seq.overworld import OverworldToMeadow, OverworldToCavern
 from evo1.seq.papurika import MeadowFight, PapurikaVillage
 from evo1.seq.start import Evoland1StartGame
+from evo1.seq.crystal_cavern import CrystalCavern
 
 __all__ = [
     "Evoland1StartGame",
@@ -9,4 +10,6 @@ __all__ = [
     "OverworldToMeadow",
     "MeadowFight",
     "PapurikaVillage",
+    "OverworldToCavern",
+    "CrystalCavern",
 ]